+++ sshd-common/src/main/java/org/apache/sshd/common/signature/AbstractSignature.java
@@ -20,6 +20,7 @@ package org.apache.sshd.common.signature;
 
 import java.nio.charset.StandardCharsets;
 import java.security.GeneralSecurityException;
+import java.security.Key;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.SignatureException;
@@ -41,7 +42,8 @@ public abstract class AbstractSignature implements Signature {
     private final String algorithm;
 
     protected AbstractSignature(String algorithm) {
-        this.algorithm = ValidateUtils.checkNotNullAndNotEmpty(algorithm, "No signature algorithm specified");
+        this.algorithm =
+            ValidateUtils.checkNotNullAndNotEmpty(algorithm, "No signature algorithm specified");
     }
 
     @Override
@@ -53,12 +55,16 @@ public abstract class AbstractSignature implements Signature {
      * Initializes the internal signature instance
      *
      * @param algo The signature's algorithm
+     * @param the {@link Key} that is provided for initialization - a {@link PrivateKey}
+     * for signing and a {@link PublicKey} for verification
      * @param forSigning If {@code true} then it is being initialized for signing,
      * otherwise for verifying a signature
      * @return The {@link java.security.Signature} instance
      * @throws GeneralSecurityException if failed to initialize
      */
-    protected java.security.Signature doInitSignature(String algo, boolean forSigning) throws GeneralSecurityException {
+    protected java.security.Signature doInitSignature(
+            String algo, Key key, boolean forSigning)
+                throws GeneralSecurityException {
         return SecurityUtils.getSignature(algo);
     }
 
@@ -73,27 +79,31 @@ public abstract class AbstractSignature implements Signature {
 
     @Override
     public byte[] sign() throws Exception {
-        java.security.Signature signature = Objects.requireNonNull(getSignature(), "Signature not initialized");
+        java.security.Signature signature =
+            Objects.requireNonNull(getSignature(), "Signature not initialized");
         return signature.sign();
     }
 
     @Override
     public void initVerifier(PublicKey key) throws Exception {
         String algo = getAlgorithm();
-        signatureInstance = Objects.requireNonNull(doInitSignature(algo, false), "No signature instance create");
+        signatureInstance = Objects.requireNonNull(
+            doInitSignature(algo, key, false), "No signature instance create");
         signatureInstance.initVerify(Objects.requireNonNull(key, "No public key provided"));
     }
 
     @Override
     public void initSigner(PrivateKey key) throws Exception {
         String algo = getAlgorithm();
-        signatureInstance = Objects.requireNonNull(doInitSignature(algo, true), "No signature instance create");
+        signatureInstance = Objects.requireNonNull(
+            doInitSignature(algo, key, true), "No signature instance create");
         signatureInstance.initSign(Objects.requireNonNull(key, "No private key provided"));
     }
 
     @Override
     public void update(byte[] hash, int off, int len) throws Exception {
-        java.security.Signature signature = Objects.requireNonNull(getSignature(), "Signature not initialized");
+        java.security.Signature signature =
+            Objects.requireNonNull(getSignature(), "Signature not initialized");
         signature.update(hash, off, len);
     }
 
@@ -138,7 +148,8 @@ public abstract class AbstractSignature implements Signature {
     }
 
     protected boolean doVerify(byte[] data) throws SignatureException {
-        java.security.Signature signature = Objects.requireNonNull(getSignature(), "Signature not initialized");
+        java.security.Signature signature =
+            Objects.requireNonNull(getSignature(), "Signature not initialized");
         return signature.verify(data);
     }